appointment

+++ admin-service/src/main/java/com/ust/admin/domain/Doctor.java
@@ -17,7 +17,7 @@ import javax.persistence.Table;
 public class Doctor {
 
     @Id
-    private int doctorId;
+    private Long doctorId;
     private String doctorName;
     private String department;
