+++ admin-service/src/main/java/com/ust/admin/dto/DoctorDto.java
@@ -1,4 +1,4 @@
 package com.ust.admin.dto;
 
-public record DoctorDto(int doctorId,String doctorName,String department) {
+public record DoctorDto(Long doctorId,String doctorName,String department) {
 }
